Extracted sortCriteria into its own file
